(#46) dns remove confirmation

+++ cmd/dns_delete.go
@@ -15,6 +15,18 @@ var dnsDeleteCmd = &cobra.Command{
 		if len(args) < 1 {
 			return cmd.Usage()
 		}
+
+		force, err := cmd.Flags().GetBool("force")
+		if err != nil {
+			return err
+		}
+
+		if !force {
+			if !askQuestion(fmt.Sprintf("Are you sure you want to delete %q domain", args[0])) {
+				return nil
+			}
+		}
+
 		if err := deleteDomain(args[0]); err != nil {
 			return err
 		}
@@ -30,4 +42,5 @@ func deleteDomain(domainName string) error {
 
 func init() {
 	dnsCmd.AddCommand(dnsDeleteCmd)
+	dnsDeleteCmd.Flags().BoolP("force", "f", false, "Attempt to delete without prompting for confirmation")
 }
